+++ rust/token-metadata/test/src/main.rs
@@ -561,8 +561,8 @@ fn create_metadata_account_call(
     let accounts = client.get_program_accounts(&program_key).unwrap();
     println!("--> Saved hero accounts: {}", accounts.len());
     let id = accounts.len() as u8 + 1;//app_matches.value_of("id").unwrap().parse::<u8>().unwrap();
-    let last_price = 0 as u16;
-    let listed_price = (app_matches.value_of("listed_price").unwrap().parse::<f64>().unwrap() * DEFAULT_LAMPORTS_PER_SOL as f64).round() as u16;
+    let last_price = 0 as u64;
+    let listed_price = (app_matches.value_of("listed_price").unwrap().parse::<f64>().unwrap() * DEFAULT_LAMPORTS_PER_SOL as f64).round() as u64;
     let name = app_matches.value_of("name").unwrap().to_owned();
     // let symbol = app_matches.value_of("symbol").unwrap().to_owned();
     let uri = app_matches.value_of("uri").unwrap().to_owned();
@@ -628,7 +628,7 @@ fn update_metadata_account_call(
     println!("--->Program_id: {}\n", program_key);
 
     let id = app_matches.value_of("id").unwrap().parse::<u8>().unwrap();
-    let listed_price = (app_matches.value_of("listed_price").unwrap().parse::<f64>().unwrap() * DEFAULT_LAMPORTS_PER_SOL as f64).round() as u16;
+    let listed_price = (app_matches.value_of("listed_price").unwrap().parse::<f64>().unwrap() * DEFAULT_LAMPORTS_PER_SOL as f64).round() as u64;
 
     println!("--->\n Id: {},\n Price: {}",id, listed_price);
     
@@ -761,7 +761,7 @@ fn purchase_hero_call(
 
     let id = app_matches.value_of("id").unwrap().parse::<u8>().unwrap();
     let listed_price = match app_matches.value_of("listed_price") {
-        Some(_val) => Some((app_matches.value_of("listed_price").unwrap().parse::<f64>().unwrap() * DEFAULT_LAMPORTS_PER_SOL as f64).round() as u16),
+        Some(_val) => Some((app_matches.value_of("listed_price").unwrap().parse::<f64>().unwrap() * DEFAULT_LAMPORTS_PER_SOL as f64).round() as u64),
         None => None,
     };
     let uri = match app_matches.value_of("uri") {
